requirements.txt
# 基础依赖
python-dotenv==1.0.1
Pillow==11.2.1
mss==10.0.0
numpy==1.26.4
pynput==1.7.6
sounddevice==0.4.6
soundfile==0.12.1
PySide6==6.9.1

# Google GenAI (参考code_sample/gemini.py)
google-genai==1.20.0

pyautogui
